[banshee] revert GTK3-related commits
This is the 'stable-2.6' branch of bockbuild, which means it
shouldn't include GTK3 but GTK2. We revert the patch that changes
this and the subsequent patches after that, which are:

[Updated FreeDesktop download link & Updated the pkg-config version to…]
d21c3bf

[Post-Hackfest bumps]
e769cd3

[Gtk3 compiles]
9dea3e5

These commits are not actually present in the master branch of
bockbuild as is, as they have been superseeded by changes done by
Xamarin, Stephan Sundermann and David Nielsen.