Revert to tip of banshee-2.6-stable branch of dynalon's bockbuild
This stable-2.6 branch was created on top of dynalon's bockbuild's
master branch, which had many changes from when the last 2.6.x
banshee Mac bundle was created. Trying to create a bundle from
this resulted in a banshee with many problems and glitches, so
let's go back to the distant past, which still works (modulo
some adjustments, which will be added in the following commits).